+++ s-vhs.sh
@@ -24,7 +24,7 @@ set -euo pipefail
 
 
 svhs_version() {
-    printf '%s\n' '0.4.2'
+    printf '%s\n' '0.5.0'
 }